+++ src/DotNet.Testcontainers/Core/Container/IDockerContainer.cs
@@ -1,9 +1,11 @@
-namespace DotNet.Testcontainers.Containers
+namespace DotNet.Testcontainers.Core.Container
 {
   using System;
 
   public interface IDockerContainer : IDisposable
   {
+    string Id { get; }
+
     string Name { get; }
 
     void Start();